Margin Padding Part 2
Actually interesting to go through the whole box-model again.
I remember that I skipped through most of that in the past
and then having all sorts of issues with it when later actually
trying to code a webpage. So it's smart to finally go through it,
tinker with it and learn about margin-collapse, auto-padding and
the like.
Took me several tries to finally get it fully right