Flexbox IV - Space img and info
At first I didn't add a specified gap, which worked,
but still looked off. Added the space and agree that it
looks more balanced overall.
Don't necessarily agree with the max-width. Experimented
with that before looking at the solution and felt like it
didn't really add anything to the style. Having the info
in one line is not looking off, but might depend on
overall content of the page you're trying to build.